# internal/response/type.go
package response

// JsonResponse is the standard JSON envelope for all API responses.
type JsonResponse struct {
	Message string      `json:"message,omitempty"`
	Data    interface{} `json:"data,omitempty"`
	Error   interface{} `json:"error,omitempty"`
}